+++ lscm2/src/indexed_face_set/indexed_face_set_renderable_mesh.h
@@ -47,11 +47,11 @@ namespace lscm
             auto index_count = 3 * static_cast<uint32_t> (mesh->m_faces.size());
 
             return std::make_shared<renderable_mesh>(
-                d3d11::create_immutable_vertex_buffer(device, &positions[0], positions.size() * sizeof(math::half)),
-                d3d11::create_immutable_index_buffer(device, &mesh->m_faces[0], mesh->m_faces.size() * sizeof(mesh::face)),
-                4 * sizeof(math::half),
-                vertex_count,
-                index_count
+                d3d11::create_immutable_vertex_buffer(device, &positions[0], positions.size() * sizeof(math::half))
+                , d3d11::create_immutable_index_buffer(device, &mesh->m_faces[0], mesh->m_faces.size() * sizeof(mesh::face))
+                , static_cast<uint32_t>( 4 * sizeof(math::half) )
+                , vertex_count
+                , index_count
                 );
         }
     }
